+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+## 0.0.4
+* Added keyboardType parameter for the inputBox widget
+* Added textInputAction parameter for the inputBox widget
+* Added onSubmitted callback for the inputBox widget
+
 ## 0.0.3
 * Added an intersect method on lists
 * Added a widget library
 * Added a extensions library
 
 ## 0.0.2
-* Added an input box widget
+* Added an inputBox widget
 
 ## 0.0.1
 * Sort lists of strings by using extensions